QUEUE-144: Keep Queue context counts as int

+++ src/main/java/net/openhft/chronicle/queue/impl/single/StoreTailer.java
@@ -1815,7 +1815,7 @@ class StoreTailer extends AbstractCloseable
         }
 
         @Override
-        public long contextCount() {
+        public int contextCount() {
             return queue.rollCycle().toCycle(index());
         }
